+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - Allow import java classes via code-actions `Add import 'x.y.Z'`.
   - Make `add-import` command/code-action smarter, checking if there is already a package import to just include the classname instead of full import.
   - Add new refactoring `cycle-keyword-auto-resolve`, e.g. `::bar` -> `:foo/bar` and vice-versa. #1128
+  - Supporting find definition/references/implementations of `definterface`.
 
 - API/CLI
   - Allow specify analysis type for `dump` command, default to analyze project only analysis. #1383